0.0.8.1.10
Former-commit-id: 6394b393a3d6d86263e96ce83723535e17986f69 [formerly 5580ce563d1e1b5967d2e03d5c195618594f2b81]
Former-commit-id: 4c80cb2690dc968b668341a9914fbd2aaffaa2de

+++ api/Cache.php
@@ -31,7 +31,7 @@ class Cache extends Simpla
 		self::$config = array_merge($config, $ini_config);
 		
 		//меняем систему счисления, чтобы chmod и mkdir правильно обрабатывали права, заданные в виде строки
-		self::$config['default_chmod'] = octdec(self::$config['default_chmod']);
+		self::$config['default_chmod'] = octdec((int)self::$config['default_chmod']);
 	}
